requirements.txt
Django==1.11.15
django-adminlte2==0.4.1
django-crequest==2018.5.11
django-environ==0.4.5
django-formset-js-improved==0.5.0.2
django-jquery-js==3.1.1
django-model-utils==3.1.2
django-widget-tweaks==1.4.3
environ==1.0
Pillow==5.4.1
psycopg2-binary==2.7.7
pytz==2018.9
